hive/requirements.txt
pyspark==3.0.2 # version must be equal spark version (check by command: spark-submit --version)
PyYAML>=6.0